Move platform/permissions to infra/auth
Change-type: patch
